fix: reuse function to detect right aws partition and support gov too
